Added new magnet functionality and help
This commit adds a new bit of magnet functionality which
shows where the magnets are without needing to get the field
at all points, which makes this a lot fatser. To accomplish the
magnet map, I re-used some code from a Geant4 project which
produces a nice 2d histogram of whatever I want.

Also, I've revamped the help given by the program, and added a
few new options for the config file. See the help message for full
details